feat(datasets): toy world (geometric emission adapters), jsonl loader, bundled sample corpus
Toy world ports the prototype's 8-component x 2-sub-mode geometry with
skewed demand; ToyChatModel/ToyJudge/PassthroughEmbedder implement the
prompt-parsing contract (toychunk/toyquery tokens, REVISE_REQUEST
tightening). Sample corpus: deterministic 200-chunk/120-query fintech
docs, byte-stable regeneration guard (PLAN D18).

+++ src/ragsynth/steps/curator.py
@@ -93,7 +93,7 @@ class Curator(PipelineStep):
         keep: set[str] = set()
         for key, frac in self.target_mix.items():
             pool = by_stratum.get(key, [])
-            n_keep = int(round(total * frac))
+            n_keep = round(total * frac)
             picks = rng.choice(len(pool), size=min(n_keep, len(pool)), replace=False)
             keep.update(pool[int(i)].record_id for i in picks)
         return [r for r in records if r.record_id in keep]
